% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantage.R
\name{alphavantage}
\alias{alphavantage}
\title{Retrieve real-time data from AlphaVantage}
\usage{
alphavantage(sym)
}
\arguments{
\item{sym}{Character value for the ticker}
}
\value{
A data.table object
}
\description{
Fetch a real-time market data series from AlphaVantage
}
\details{
Several optional parameters could be set, but are not currently.
}
\author{
Dirk Eddelbuettel
}
