% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assignFormals}
\alias{assignFormals}
\title{Assign formal arguments from function}
\usage{
assignFormals(f)
}
\arguments{
\item{f}{A function}
}
\value{
Nothing, but a side effect of assignment in global
 environment
}
\description{
Debugging helper to assign formals from function
}
\details{
(Beta) Attempt get default values from a given function,
extracting its arguments and assigning which should help in
debugging via \code{browser()} and other helper functions.
}
\author{
Dirk Eddelbuettel
}
