% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCRANStatus.R
\name{checkCRANStatus}
\alias{checkCRANStatus}
\title{Maintainer Status at CRAN}
\usage{
checkCRANStatus(email, cache, cache.life = 24 * 3600)
}
\arguments{
\item{email}{A character variable with the maintainer email}

\item{cache}{A character variable with an optional cache file,
default value is to use a file inside the per-session
temporary directory}

\item{cache.life}{A numeric timeout, defaults to one day}
}
\value{
Nothing, the sideffect of the display is the main effect
}
\description{
Report Maintainer Status at CRAN
}
\details{
This function retrieves the maintainer status (given an email)
at CRAN. Values are optionally cached; the default cache location
is inside the per-session temporary directory as R should not write
elsewhere.
}
\author{
Brodie Gaslam (with minor modifications by Dirk Eddelbuettel)
}
