% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNonAscii.R
\name{checkPackageAsciiCode}
\alias{checkPackageAsciiCode}
\title{checkPackageAsciiCode}
\usage{
checkPackageAsciiCode(dir, respect_quotes = FALSE)
}
\arguments{
\item{dir}{Character variable with path to directory to be checked}

\item{respect_quotes}{Logical variable whether quotes need to be checked}
}
\value{
A vector of things that are wrong per this function, also displayed
on standard output
}
\description{
Check a package directory for non-ASCII characters in source files.
}
\details{
The function is a renamed and slightly edited copy of the base R function
\code{.check_package_ASCII_code}. It uses an unexported C function, also
included, from base R, called as \code{_check_nonASCII}.
}
\examples{
\dontrun{
checkPackageAsciiCode(".", FALSE)
}
}
\author{
Dirk Eddelbuettel for this version, R Core for the underlying code
}
