% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranORCIDs.R
\name{cranORCIDs}
\alias{cranORCIDs}
\title{Get ORCID IDs known at CRAN}
\usage{
cranORCIDs()
}
\value{
A data frame with three columns given (ie first name), family (name) and oid.
}
\description{
This function returns a three-column data frame with first name, family name and
orcid for all maintainers with (optional) ORCID fields at CRAN.
}
\seealso{
A new function \code{tools::CRAN_authors_db()} in r-devel as of August 2024.
}
\author{
Kurt Hornik (plus small tweaks by Dirk Eddelbuettel)
}
