% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dang.R
\name{dang}
\alias{dang}
\title{Execute the last command with a new function, but the same arguments.}
\usage{
dang(new_fn)
}
\arguments{
\item{new_fn}{Raw new function with which to evaluate the old arguments.}
}
\description{
You mistyped it, but don't want to up-arrow, ctrl+a, delete. dang!
}
\examples{
> pritn("Hello world!")
# Error: could not find function "pritn"
> dang(print)
# [1] "Hello world!"
}

