% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotivate.R
\name{demotivate}
\alias{demotivate}
\title{Demotivating quote}
\usage{
demotivate(x, width = NULL)
}
\arguments{
\item{x}{A value between 0 and 1 chosen to select a quote; if not given a random
uniform draw is made.}

\item{width}{The desired display width.}
}
\value{
The formatted is returned invisibly.
}
\description{
Display a demotivating quote
}
\details{
This function is port of the 'Stata' ado file at
\url{http://fmwww.bc.edu/repec/bocode/d/demotivate.ado}
}
\examples{
demotivate()
}
\seealso{
\url{http://fmwww.bc.edu/repec/bocode/d/demotivate.ado}
}
\author{
Dirk Eddelbuettel for this function, and Kevin Denny for the 'Stata' original.
}
