% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inGit.R
\name{getGitRoot}
\alias{getGitRoot}
\alias{inGit}
\title{Are we in a \code{git} repository?}
\usage{
getGitRoot(cwd = getwd())

inGit(cwd = getwd())
}
\arguments{
\item{cwd}{The start directory, default to the current working
directory}
}
\value{
For \code{getGitRoot()}, the path of the directory containing
 the \code{.git} directory, ie the project root directory, or an
 empty string in case the search started outside a \code{git} directory.
 The \code{inGit()} function returns a boolean as to whether a \code{git}
 repository was found or not.
}
\description{
The \code{getGitRoot()} function recursively ascends the
filesystem tree from the given directory until it either finds a
directory \code{.git}, or the top-level directory to abort the
search. The root directory of the \code{git} repository is
returned, with an exmpty string in the case of no repository.  The
\code{inGit()} function turns this into boolean predicate returning
either \code{TRUE} or \code{FALSE}.
}
\examples{
inGit()
}
\author{
Dirk Eddelbuettel
}
