% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inGit.R
\name{inGit}
\alias{inGit}
\title{Are we in a \code{git} repository?}
\usage{
inGit(cwd = getwd())
}
\arguments{
\item{cwd}{The start directory, default to the current working
directory}
}
\value{
A boolean value indicating whether or not the given
directory is, in fact, in a \code{git} repository.
}
\description{
This function recursively ascends the filesystem tree from the
given directory until it either finds a directory \code{.git}, or
the top-level directory to abort the search.
}
\details{
On Windows, \code{FALSE} is returned unconditionally.
}
\examples{
inGit()
}
\author{
Dirk Eddelbuettel
}
