% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConnected.R
\name{isConnected}
\alias{isConnected}
\title{Is the current session (networked) and connected?}
\usage{
isConnected(site = "https://www.google.com")
}
\arguments{
\item{site}{Character variable with site to try to connect to,
defauls to \code{http://www.google.com}}
}
\value{
A boolean value indicating networking status
}
\description{
Function to (heuristically) test for a network connection by attempting
to connect to a given website.
}
\details{
The main page of Google is used as a proxy for overall network
connectivity as Google is generally 'network-close' and the page
is relatively small.  If a network is unavailable this fails
generally already on domain name service resolution. Special thanks
to Barry and Brodie for a very helpful discussion re-discovering this
function.
}
\author{
Dirk Eddelbuettel
}
