% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{limitDataTableCores}
\alias{limitDataTableCores}
\title{Set data.table threads respecting default settingss}
\usage{
limitDataTableCores(ncores, verbose = FALSE)
}
\arguments{
\item{ncores}{A numeric or character variable with the desired
count of threads to use}

\item{verbose}{A logical value with a default of \sQuote{FALSE} to
operate more verbosely}
}
\value{
The return value of the \pkg{data.table} function
\code{setDTthreads} which is called as a side-effect.
}
\description{
Set threads for data.table respecting possible local settings
}
\details{
This function set the number of threads \pkg{data.table} will use
while reflecting two possible machine-specific settings from the
environment variable \sQuote{OMP_THREAD_LIMIT} as well as the R
option \sQuote{Ncpus} (uses e.g. for parallel builds).
}
\author{
Dirk Eddelbuettel
}
