% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ls.objects}
\alias{ls.objects}
\alias{lsos}
\alias{showMemoryUse}
\title{List largest objects and show memory use}
\usage{
ls.objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)

lsos(..., n = 10)

showMemoryUse(sort = "size", decreasing = FALSE, limit)
}
\arguments{
\item{pos}{Position in search path, defaults to 1}

\item{pattern}{Pattern argument pass on to \code{ls}}

\item{order.by}{Optional sort order column}

\item{decreasing}{Optional switch for decreasing or increasing
sort order}

\item{head}{Optional switch to show \code{head(..., n)}}

\item{n}{Number of elements to show, default to 5}

\item{...}{Passed through from \code{lsos} to \code{ls.objects}}

\item{sort}{Sort columns, defauls to \sQuote{size}}

\item{limit}{Optional cap on displayed number of items}
}
\value{
The displayed data.frame is returned to, the main purpose
effect however is the displayed information
}
\description{
List object and memory used by these objects
}
\details{
These helper functions have evolved over the years; some were also
posted on StackOverflow in response to
\url{https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}
\author{
Dirk Eddelbuettel, based on and extending code in mailing list
posts by Petr Pikal and David Hinds
}
