% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivate.R
\name{motivate}
\alias{motivate}
\title{Display a motivating quote}
\usage{
motivate(x, width = NULL)
}
\arguments{
\item{x}{A index value chosen to select a quote; if not given a quote is
chosen randomly.}

\item{width}{The desired display width.}
}
\value{
The formatted is returned invisibly.
}
\description{
Display a motivating quote for users.
}
\details{
This function is a port of \code{motivate} command for 'Stata' from the
ado file at \url{http://fmwww.bc.edu/repec/bocode/m/motivate.ado}
}
\examples{
motivate()
}
\seealso{
\url{https://ideas.repec.org/c/boc/bocode/s458565.html}
}
\author{
Dirk Eddelbuettel for this function, and Kabira Namit for the
'Stata' original.
}
