% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dang.R
\name{new_named_arguments}
\alias{new_named_arguments}
\title{Replace existing or insert new named arguments into a call.}
\usage{
new_named_arguments(call, args)
}
\description{
Replace existing or insert new named arguments into a call.
}

