% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.xts}
\alias{print.xts}
\title{Printing helper function for xts object}
\usage{
\method{print}{xts}(x, ..., n = 10)
}
\arguments{
\item{x}{An xts objects}

\item{...}{Ignored, but needed for consistency with \code{print}
generic.}

\item{n}{The default number of lines to display, default is 10.}
}
\value{
The return from \code{print}
}
\description{
Print helper function for xts objects.
}
\author{
Dirk Eddelbuettel
}
