% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dang.R
\name{replace_function_name}
\alias{replace_function_name}
\title{Replace the top level function name of a call.}
\usage{
replace_function_name(call, new_fn)
}
\description{
Replace the top level function name of a call.
}

