% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dang.R
\name{retrieve_last_command}
\alias{retrieve_last_command}
\title{Grab the second most recent command in the history}
\usage{
retrieve_last_command()
}
\description{
Grab the second most recent command in the history
}

