% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollMinMax.R
\name{rollMinMax}
\alias{rollMinMax}
\title{Rolling Minimum or Maximum over a Fixed Window}
\usage{
rollMinMax(x, window, min = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{window}{An interger with the size of the rolling window.}

\item{min}{A logical which, if true, selects minimum, else maximum}
}
\value{
A vector of the same length as 'x' with elements 1 to
(length(window) - 1) containing NAs.
}
\description{
This implementation is minimal without error checking, or NA handling.
It is taken from the \pkg{ichimoku} package which had several more
complicated variants, and is reused here with just \pkg{tidyCpp}.
}
\details{
The \pkg{ichimoku} variant is by Charlie Gao and credits Andrew Uhl for
the initial implementation.
}
\author{
Dirk Eddelbuettel for this version, R Core for the underlying code
}
