% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowedPackages.R
\name{shadowedPackages}
\alias{shadowedPackages}
\title{Find Shadowed Packages}
\usage{
shadowedPackages()
}
\value{
data.table object with packages that are shadowed
}
\description{
Return a data.table object with \sQuote{shadowed} packages,
meaning package which are installed in more than directory on the
\code{.libPaths}.
}
\details{
The function relies on the base R functions
\code{installed.packages()} to find all packages, as well as the
base R abilility to compare version numbers (once properly
converted to \code{package_version} type).
}
\author{
Dirk Eddelbuettel
}
