% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{silent}
\alias{silent}
\title{Silently load a library}
\usage{
silent(...)
}
\arguments{
\item{...}{Passed though}
}
\value{
Nothing, but the desired library is attached
}
\description{
Silently attach a library to the search path.
}
\details{
This function wraps \code{suppressMessages} around the call to
\code{library}
}
\author{
Dirk Eddelbuettel
}
