% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{ymd}
\alias{ymd}
\title{Format a Date(time) object as ymd}
\usage{
ymd(pt)
}
\arguments{
\item{pt}{A \code{POSIXt} Datetime or a \code{Date} object}
}
\value{
A character object formatted as \sQuote{YYYYMMDD}
}
\description{
Format a Date(time) object as ymd
}
\examples{
ymd(anytime("2016-09-01 10:11:12.123456"))
ymd(anydate("2016-Sep-01"))
}
\author{
Dirk Eddelbuettel
}
