% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{as.data.table.xts}
\alias{as.data.table.xts}
\title{Convert an xts object into a data.table}
\usage{
as.data.table.xts(x)
}
\arguments{
\item{x}{An xts object}
}
\value{
A data.table object with new columns date, time, micros,
 pt providing, respectively the data as \code{IDate}, time as
\code{ITime}, microseconds (rounded) and numeric
\code{POSIXct}.
}
\description{
Convert an xts object into a data.table
}
\details{
This is still experimental.  Note that all four \emph{added}
columns are keys to the data.table object, and that
\code{setNumericRounding(0)} is executed too,
}
\author{
Dirk Eddelbuettel
}
