% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dang.R
\name{dang}
\alias{dang}
\title{Fix errors in the last command in the history.}
\usage{
dang(...)
}
\arguments{
\item{...}{Either a single unnamed argument or a series of named arguments.}
}
\description{
Call without one unnamed argument to replace the old calling function with
the new argument. Call with one or more named arguments to replace or
inject new arguments into the last call.
}
\examples{
pritn("Hello world!")
# Error: could not find function "pritn"
dang(print)

sample(letters, size = 26, replace = FALSE)
dang(replace = TRUE)
}

