% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demotivate.R
\name{demotivate}
\alias{demotivate}
\title{Display a demotivating quote}
\usage{
demotivate(x, width = NULL)
}
\arguments{
\item{x}{A index value chosen to select a quote; if not given a quote is
chosen randomly.}

\item{width}{The desired display width.}
}
\value{
The formatted is returned invisibly.
}
\description{
Display a demotivating quote to remind the users of the harsh reality
econometric and statistical practice.
}
\details{
This function is a port of \code{demotivate} command for 'Stata' from the
ado file at \url{http://fmwww.bc.edu/repec/bocode/d/demotivate.ado}.
}
\examples{
demotivate()
}
\seealso{
\url{https://ideas.repec.org/c/boc/bocode/s458576.html}
}
\author{
Dirk Eddelbuettel for this function, and Kevin Denny for the
'Stata' original.
}
