% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderMicrobenchmarkResults.R
\name{reorderMicrobenchmarkResults}
\alias{reorderMicrobenchmarkResults}
\alias{rmr}
\title{Reorder microbenchmark Results}
\usage{
reorderMicrobenchmarkResults(res, ordervar = "median")

rmr(res, ordervar = "median")
}
\arguments{
\item{res}{An object returned from `microbenchmark::microbenchmark`}

\item{ordervar}{An option character variable selecting a column to order on,
defaults to 'median'}
}
\value{
A modified version of `res` with an ordering by `order`
}
\description{
Reorder microbenchmark results
}
\details{
The lovely `microbenchmark` package returns its result is a simple structure
with `factor` variable which it leaves unordered.  The printed as well as
plotted results become a little more expressive if an order is added, which
this helper function does.

'rmr' can be used as a shorter alias
}
